+++ src/NEventStore.Persistence.MySql.Tests/PersistenceEngineFixture.cs
@@ -11,7 +11,7 @@ namespace NEventStore.Persistence.AcceptanceTests
     {
         public PersistenceEngineFixture()
         {
-#if !NETSTANDARD2_0
+#if NET461
             _createPersistence = pageSize =>
                     new SqlPersistenceFactory(
                         new EnviromentConnectionFactory("MySql", "MySql.Data.MySqlClient"),