+++ src/NEventStore.Persistence.PostgreSql.Tests/PersistenceEngineFixture.cs
@@ -10,7 +10,7 @@ namespace NEventStore.Persistence.AcceptanceTests
     {
         public PersistenceEngineFixture()
         {
-#if !NETSTANDARD2_0
+#if NET461
             _createPersistence = pageSize =>
                 new SqlPersistenceFactory(
                     new EnviromentConnectionFactory("PostgreSql", "Npgsql"),