+++ src/NEventStore.Persistence.Sql/SnapshotExtensions.cs
@@ -1,6 +1,7 @@
 namespace NEventStore.Persistence.Sql
 {
     using System.Data;
+    using Microsoft.Extensions.Logging;
     using NEventStore.Logging;
     using NEventStore.Serialization;
 
@@ -9,11 +10,11 @@ namespace NEventStore.Persistence.Sql
         private const int BucketIdIndex = 0;
         private const int StreamRevisionIndex = 2;
         private const int PayloadIndex = 3;
-        private static readonly ILog Logger = LogFactory.BuildLogger(typeof (SnapshotExtensions));
+        private static readonly ILogger Logger = LogFactory.BuildLogger(typeof (SnapshotExtensions));
 
         public static Snapshot GetSnapshot(this IDataRecord record, ISerialize serializer, string streamIdOriginal)
         {
-            if (Logger.IsVerboseEnabled) Logger.Verbose(Messages.DeserializingSnapshot);
+            Logger.LogTrace(Messages.DeserializingSnapshot);
 
             return new Snapshot(
                 record[BucketIdIndex].ToString(),