+++ src/NEventStore.Persistence.Sql/SqlDialects/SqliteDialect.cs
@@ -2,6 +2,9 @@ namespace NEventStore.Persistence.Sql.SqlDialects
 {
     using System;
 
+    /// <summary>
+    /// Dialect that should be used with System.Data.Sqlite
+    /// </summary>
     public class SqliteDialect : CommonSqlDialect
     {
         public override string InitializeStorage
@@ -24,24 +27,29 @@ namespace NEventStore.Persistence.Sql.SqlDialects
         {
             string message = exception.Message.ToUpperInvariant();
             return message.Contains("DUPLICATE") || message.Contains("UNIQUE") || message.Contains("CONSTRAINT");
-
         }
 
         public override DateTime ToDateTime(object value)
         {
-#if !NETSTANDARD2_0
-            return ((DateTime) value).ToUniversalTime();
-#else
+            return ((DateTime)value).ToUniversalTime();
+        }
+    }
+
+    /// <summary>
+    /// Dialect that should be used with Microsoft.Data.Sqlite.
+    /// </summary>
+    public class MicrosoftDataSqliteDialect : SqliteDialect
+    {
+        public override DateTime ToDateTime(object value)
+        {
             // original code
             // return ((DateTime) value).ToUniversalTime();
-
             // not working, 'value' is already an utc value in ISO86001 format
             // Convert.ToDateTime() will return an unspecified kind
             // return Convert.ToDateTime(value).ToUniversalTime();
 
             // CommitStap is always an UTC value
             return DateTime.SpecifyKind(Convert.ToDateTime(value), DateTimeKind.Utc);
-#endif
         }
     }
 }